generalize collection iterator
